task: add c++ source files and meson.build
The meson.build file is adapted from the mohair repository, and the
source files are a selection of source files from the mohair repository.

The source files in `src/cpp/mohair` and `src/cpp/substrait` are
generated using the buf.build tool.

This repository should build a minimal library that makes it easy to
link against substrait (and my own mohair protocol that is an additive
extension) so that it's possible to finally avoid linking against other
libraries that use substrait and coming up to errors regarding duplicate
protobuf descriptors.