task: split PlanAnchor into SuperPlan and SubPlan
This changes the PlanAnchor message so that we don't conceptually
conflict with the new plan anchor reference being proposed to substrait.
This also sets up some new attributes to eventually leverage anchors in
SuperPlan and adds a new SubPlan message that can be used to refer to a
sub plan.