task: code reorganization
Changed the namespace of the code from `mohair` to `skytether` and
changed the namespace of the protocol sources from `skytether` to
`skyproto`. This should be a bit clearer for protocol sources and also
allow for me to collapse `mohair` and `skytether` namespaces into one.

All of this prepares me to integrate better with query engines (e.g.
duckdb) for code that splits and merges query plans