task: added some query processing
Simple query processing was originally in the mohair repository, but
moving it into this library allows others to be able to use it, but also
moves it into a place where query engines can use it.

The mohair repository (skytether library) contains code that should be
an abstraction above query engines, but when execution of a query plan
should also split the query plan, it is better for the code to be
visible to the query engine. This allows the skytether library to
delegate execution responsibility to the query engine without having to
be deeply integrated (or at least any deeper than necessary).

This also makes some translation code from substrait to an engine's
internal plan representation potentially re-usable.