task: extended SystemPlan and refactored functions
SystemPlan was modified to take ownership of a substrait `Plan` it is
constructed from and various functions have been renamed
(`MohairPlanFrom` -> `SystemPlanFrom`), refactored (take unique_ptr
instead of reference), and otherwise minorly altered to reflect a
variety of nuanced changes.

The first main change is that the mohair library should totally own the
SystemPlan and its construction. Then, it should be an intermediate step
to constructing an EnginePlan, which should be owned by skytether.
Just like DuckDB only knows its own representation, a SystemPlan is in
its own domain; however, skytether will bridge EnginePlans (adapter for
a particular engine's representation such as DuckDB) and the SystemPlan,
including having its own translation from a SystemPlan to an EnginePlan
(even if that translation is primarily vendored).