task: a bit of re-organization and fixed build
The re-organization was to accommodate some extra changes to make the
skytether library still work with this library after all of the code
porting.

The build fix was due to adding build files to subdirectories, but when
header files are added to a list to provide to `install_headers`, all of
the files were installed to the same directory instead of in the
appropriate subdirectory. So, now each meson.build file that specifies
header files installs them to the appopriate destination directory.