task: added splitting and merging of plans
As I added more functions for moving objects around the plan (e.g.
moving from one `Rel` to another `Rel`), I decided to add some naive
implementations to `mohair.hpp` and `mohair.cpp`. These are meant to be
implementations that interact with substrait directly and so don't need
to be in the operators or plans files.

Then, a lot of logic was added to `PlanSplit` with some supporting
changes made to various operators. Pipelines are not yet rebuilt after
merging plans, so they have to be manually rebuilt (another call to
`BuildPipelines`).

Finally, I added a new program, `decompose-substrait`, that just checks
the splitting and merging code is functional. It is likely to become the
basis of more tests but also of an interface to cooperative
decomposition (without execution).