task: Updated decompose-substrait to be complete
decompose-substrait now specifies a parameter (`count_splits`) and then
recursively calls `ProcessPlanAsService` to mimic single-site
cooperative decomposition. This isn't yet a thorough test, but it
provides a quick check that the code works for a test case.

A comparison to "Iterative dynamic programming" is necessary and should
heavily explore the difference between cooperative decomposition in
local vs distributed contexts.