task: added schema resolution
This commit primarily adds schema resolution at the time that a
substrait plan is internalized (`MohairFrom` function). The purpose of
schema resolution is to ensure that when splitting a plan certain
attributes are preserved during optimization.

This initially was implemented to address a bug that turned out to be a
different issue. However, knowing the output schema of each operator is
useful for debugging and will allow an optimizer to eliminate
unnecessary attributes from subplans.