task: parameterized DecomposeAlg option
DecomposeAlg was a cpp enum that specified a strategy for choosing a
candidate split in a query plan. It has now been moved into the mohair
protocol so that a remote client or service can specify it.

Additionally, I updated `read-substrait` to check if the provided plan
file path exists (it used to segfault, which seemed like bad UX)