task: added log initialization with specified name
Having only a timestamp makes it more annoying to associate a log file
with the correct service. Also, I needed a log initialization path to
ensure the log is created at a useful time (late initialization without
a name made it impossible to distinguish between services running on the
same server in the same location).