refactoring: progress on adopting templates
This commit contains the bulk of transitioning from `MohairOp` which
uses vtables and polymorphism for op-specific logic to `SubstraitOp`
which is an interface to the templated `SubstraitOpImpl`.

SubstraitOpImpl tries to use template metaprogramming and
specializations to enable a more efficient way to interact with
substrait code without having to frequently go through switch case
statements. The use of an interface class (SubstraitOp) seems to be the
most ergonomic way to call into templated code, but hopefully it doesn't
prevent any of the benefits we get from using the templated code.

That being said, this transition largely isn't a performance improvement
(or at least I haven't hypothesized if there are any yet) except in
copying of protobuf messages. Previously, copying protobuf messages
defaulted to recursive copying and I would clear everything that was
recursively copied. This will avoid that and hopefully the
specializations for that are already implemented.

This code still needs fixing but the main difference will be that
string_view doesn't seem a great way to get names for some operators, or
I'll have to change what that looks like.