task: added operator IDs and hashes
This adds remaining support for operator IDs as well as identifiers of
pipelines and pipeline stages.

A pipeline is identified by a hash of its "flow descriptor." A flow
descriptor is a concatenation of the pipeline's sink operator, pipeline
operators, and source operator as the pipeline is built. This uses the
standard library's string hash functor.

A pipeline stage is identified by a hash of its sink operator, each
member pipeline's hash, and a hash of an "upstream stage" (upstream is
closer to a scan).

Hashes are initially computed when a `SystemPlan` is parsed (in the
function `SystemPlanFrom`) and are not re-computed if available: first
the hash is initialized from the `Plan` if available, then it is simply
returned if a non-zero value is present in memory.

I do not have the testing I should, but the intent is that only the
service receiving the source query plan from the client should calculate
the hashes and every downstream engine should re-use those hash values.
As a query plan is decomposed and propagated, some aspects of
identification are bound to change (due to sink operators being
conditionally propagated). If sink operators are omitted from hash
calculation, it may be possible to calculate hashes dynamically while
maintaining identity, but for now this is a straightforward approach.