task: updated how pipe ids are persisted
Previously, there was an error-prone approach of copying a pipeline ID
(`alias_hash`) to its children when delegating join inputs. The approach
was further complicated by how pipeline IDs were calculated (only one
branch was used due to how checking for an existing ID was done).

This commit changes the approach to let a pipeline persist an ID to the
last pipeline operator and let pipeline stages persist to the stage
sink. This avoids the need to copy the ID to its inputs, and allows the
ID of a pipeline to be based on pipeline operators and the source. To
make this approach accurate, pipelines send their own ID to their next
downstream pipeline so that when a pipeline's ID is calculated it
accommodates all upstream pipelines.